# internal/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AdminStats:
    properties:
      active_shipments:
        type: integer
      container_types:
        type: integer
      success_rate:
        type: number
      total_users:
        type: integer
    type: object
...
  dto.ContainerResponse:
    properties:
      description:
        type: string
      id:
        type: string
      inner_height_mm:
        type: number
      inner_length_mm:
        type: number
      inner_width_mm:
        type: number
      max_weight_kg:
        type: number
      name:
        type: string
    type: object
  dto.CreateContainerRequest:
    properties:
      description:
        maxLength: 500
        type: string
      inner_height_mm:
        type: number
      inner_length_mm:
        type: number
      inner_width_mm:
        type: number
      max_weight_kg:
        type: number
      name:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - inner_height_mm
    - inner_length_mm
    - inner_width_mm
    - max_weight_kg
    - name
    type: object
...
  dto.DashboardStatsResponse:
    properties:
      admin:
        $ref: '#/definitions/dto.AdminStats'
      operator:
        $ref: '#/definitions/dto.OperatorStats'
      planner:
        $ref: '#/definitions/dto.PlannerStats'
    type: object
...
  dto.OperatorStats:
    properties:
      active_loads:
        type: integer
      avg_time_per_load:
        type: string
      completed:
        type: integer
      failed_validations:
        type: integer
    type: object
...
  dto.PlannerStats:
    properties:
      avg_utilization:
        type: number
      completed_today:
        type: integer
      items_processed:
        type: integer
      pending_plans:
        type: integer
    type: object
...
  dto.UpdateContainerRequest:
    properties:
      description:
        maxLength: 500
        type: string
      inner_height_mm:
        type: number
      inner_length_mm:
        type: number
      inner_width_mm:
        type: number
      max_weight_kg:
        type: number
      name:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - inner_height_mm
    - inner_length_mm
    - inner_width_mm
    - max_weight_kg
    - name
    type: object
...
  response.APIResponse:
    properties:
      data: {}
      errors:
        items:
          $ref: '#/definitions/response.ErrorDetail'
        type: array
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
  response.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
      source: {}
    type: object
  response.Meta:
    properties:
      count:
        type: integer
      current_page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Server for Load Stuffing Calculator application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Load Stuffing Calculator API
  version: "1.0"
paths:
...
  /containers:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of containers.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ContainerResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: List containers
      tags:
      - containers
    post:
      consumes:
      - application/json
      description: Creates a new container type. Requires admin privileges.
      parameters:
      - description: Container Creation Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateContainerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ContainerResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create a new container
      tags:
      - containers
  /containers/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a container by ID. Requires admin privileges.
      parameters:
      - description: Container ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete a container
      tags:
      - containers
    get:
      consumes:
      - application/json
      description: Retrieves container details by ID.
      parameters:
      - description: Container ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ContainerResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get a container by ID
      tags:
      - containers
    put:
      consumes:
      - application/json
      description: Updates an existing container. Requires admin privileges.
      parameters:
      - description: Container ID
        in: path
        name: id
        required: true
        type: string
      - description: Container Update Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateContainerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update a container
      tags:
      - containers
  /dashboard/stats:
    get:
      consumes:
      - application/json
      description: Returns aggregated statistics for the dashboard based on user role.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.DashboardStatsResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Dashboard Stats
      tags:
      - dashboard
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"